refactor(ml-1268): extract outcome view-builders to utils.js
The outcome handler had three inline option-mapping blocks for
intermediate / terminal-single / terminal-multi, plus a side-effecting
logEmptyTextIfNeeded inside types.map. Move the pure view-model
construction (and the existing classifyOutcome / ctaLabelFor helpers)
into utils.js alongside the controller, matching the
controller.js / utils.js convention used elsewhere in the repo.
Handler is now a 6-line dispatch; the empty-text logging on terminal-
multi is a plain for-of before the view build instead of riding inside
the map. Pure-helper tests move into utils.test.js.
